Run `nix flake update`
The last commit I was on had a kernel that was broken with my WiFi card
and this finally fixed it! There was some stupid bug where
if I enabled specific VPNs it would lock up the ip subsystem.
It happened when I added dotnet as a dev tool and at that point
I felt like I had had enough so it's gone now